backend/requirements.txt
Werkzeug
pymongo
gunicorn
Flask
flask-marshmallow
marshmallow
mongomock
PyJWT
python-dotenv